+++ packages/terra-section-header/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Props table markdown file import added to doc site page
+
 
 2.22.0 - (September 3, 2019)
 ------------------